+++ pkg/azurecluster/validate_create.go
@@ -7,18 +7,22 @@ import (
 	"github.com/giantswarm/micrologger"
 	"k8s.io/api/admission/v1beta1"
 	capzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
+	"github.com/giantswarm/azure-admission-controller/pkg/generic"
 	"github.com/giantswarm/azure-admission-controller/pkg/validator"
 )
 
 type CreateValidator struct {
 	baseDomain string
+	ctrlClient client.Client
 	logger     micrologger.Logger
 }
 
 type CreateValidatorConfig struct {
 	BaseDomain string
+	CtrlClient client.Client
 	Logger     micrologger.Logger
 }
 
@@ -26,12 +30,16 @@ func NewCreateValidator(config CreateValidatorConfig) (*CreateValidator, error) 
 	if config.BaseDomain == "" {
 		return nil, microerror.Maskf(invalidConfigError, "%T.BaseDomain must not be empty", config)
 	}
+	if config.CtrlClient == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.CtrlClient must not be empty", config)
+	}
 	if config.Logger == nil {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
 	}
 
 	v := &CreateValidator{
 		baseDomain: config.BaseDomain,
+		ctrlClient: config.CtrlClient,
 		logger:     config.Logger,
 	}
 
@@ -44,7 +52,12 @@ func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return false, microerror.Maskf(errors.ParsingFailedError, "unable to parse AzureCluster CR: %v", err)
 	}
 
-	err := validateControlPlaneEndpoint(*azureClusterCR, a.baseDomain)
+	err := generic.ValidateOrganizationLabelContainsExistingOrganization(ctx, a.ctrlClient, azureClusterCR)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
+	err = validateControlPlaneEndpoint(*azureClusterCR, a.baseDomain)
 	if err != nil {
 		return false, microerror.Mask(err)
 	}