+++ pkg/azuremachine/validate_update_test.go
@@ -4,11 +4,14 @@ import (
 	"context"
 	"testing"
 
+	securityv1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/security/v1alpha1"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"k8s.io/api/admission/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
+
+	"github.com/giantswarm/azure-admission-controller/pkg/unittest"
 )
 
 func TestAzureMachineUpdateValidate(t *testing.T) {
@@ -50,12 +53,29 @@ func TestAzureMachineUpdateValidate(t *testing.T) {
 				}
 			}
 
+			ctx := context.Background()
+			fakeK8sClient := unittest.FakeK8sClient()
+			ctrlClient := fakeK8sClient.CtrlClient()
+
+			// Create default GiantSwarm organization.
+			organization := &securityv1alpha1.Organization{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "giantswarm",
+				},
+				Spec: securityv1alpha1.OrganizationSpec{},
+			}
+			err = ctrlClient.Create(ctx, organization)
+			if err != nil {
+				t.Fatal(err)
+			}
+
 			admit := &CreateValidator{
-				logger: newLogger,
+				ctrlClient: ctrlClient,
+				logger:     newLogger,
 			}
 
 			// Run admission request to validate AzureConfig updates.
-			allowed, err := admit.Validate(context.Background(), getUpdateAdmissionRequest(tc.oldAM, tc.newAM))
+			allowed, err := admit.Validate(ctx, getUpdateAdmissionRequest(tc.oldAM, tc.newAM))
 
 			// Check if the error is the expected one.
 			switch {