+++ pkg/cluster/validate_create.go
@@ -10,6 +10,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
+	"github.com/giantswarm/azure-admission-controller/pkg/generic"
 	"github.com/giantswarm/azure-admission-controller/pkg/validator"
 )
 
@@ -51,7 +52,12 @@ func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 		return false, microerror.Maskf(errors.ParsingFailedError, "unable to parse Cluster CR: %v", err)
 	}
 
-	err := validateClusterNetwork(*clusterCR, a.baseDomain)
+	err := generic.ValidateOrganizationLabelContainsExistingOrganization(ctx, a.ctrlClient, clusterCR)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
+	err = validateClusterNetwork(*clusterCR, a.baseDomain)
 	if err != nil {
 		return false, microerror.Mask(err)
 	}