+++ pkg/cluster/validate_create_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/Azure/go-autorest/autorest/to"
+	securityv1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/security/v1alpha1"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"k8s.io/api/admission/v1beta1"
@@ -13,6 +14,7 @@ import (
 	"sigs.k8s.io/cluster-api/api/v1alpha3"
 
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
+	"github.com/giantswarm/azure-admission-controller/pkg/unittest"
 )
 
 func TestClusterCreateValidate(t *testing.T) {
@@ -151,13 +153,30 @@ func TestClusterCreateValidate(t *testing.T) {
 				}
 			}
 
+			ctx := context.Background()
+			fakeK8sClient := unittest.FakeK8sClient()
+			ctrlClient := fakeK8sClient.CtrlClient()
+
+			// Create default GiantSwarm organization.
+			organization := &securityv1alpha1.Organization{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "giantswarm",
+				},
+				Spec: securityv1alpha1.OrganizationSpec{},
+			}
+			err = ctrlClient.Create(ctx, organization)
+			if err != nil {
+				t.Fatal(err)
+			}
+
 			admit := &CreateValidator{
 				baseDomain: "k8s.test.westeurope.azure.gigantic.io",
+				ctrlClient: ctrlClient,
 				logger:     newLogger,
 			}
 
 			// Run admission request to validate AzureConfig updates.
-			allowed, err := admit.Validate(context.Background(), getCreateAdmissionRequest(tc.cluster))
+			allowed, err := admit.Validate(ctx, getCreateAdmissionRequest(tc.cluster))
 
 			// Check if the error is the expected one.
 			switch {