+++ pkg/machinepool/validate_machinepool_create_test.go
@@ -7,23 +7,19 @@ import (
 
 	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2019-07-01/compute"
 	"github.com/Azure/go-autorest/autorest/to"
-	corev1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/core/v1alpha1"
-	providerv1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/provider/v1alpha1"
-	releasev1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/release/v1alpha1"
+	securityv1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/security/v1alpha1"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"k8s.io/api/admission/v1beta1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	capzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/api/v1alpha3"
 	expcapzv1alpha3 "sigs.k8s.io/cluster-api-provider-azure/exp/api/v1alpha3"
 	"sigs.k8s.io/cluster-api/api/v1alpha3"
 	capiv1alpha3 "sigs.k8s.io/cluster-api/exp/api/v1alpha3"
-	expcapiv1alpha3 "sigs.k8s.io/cluster-api/exp/api/v1alpha3"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 
 	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
+	"github.com/giantswarm/azure-admission-controller/pkg/unittest"
 )
 
 const (
@@ -177,41 +173,9 @@ func TestMachinePoolCreateValidate(t *testing.T) {
 				panic(microerror.JSON(err))
 			}
 
-			scheme := runtime.NewScheme()
-			err = v1.AddToScheme(scheme)
-			if err != nil {
-				panic(err)
-			}
-			err = corev1alpha1.AddToScheme(scheme)
-			if err != nil {
-				panic(err)
-			}
-			err = expcapiv1alpha3.AddToScheme(scheme)
-			if err != nil {
-				panic(err)
-			}
-			err = expcapzv1alpha3.AddToScheme(scheme)
-			if err != nil {
-				panic(err)
-			}
-			err = capiv1alpha3.AddToScheme(scheme)
-			if err != nil {
-				panic(err)
-			}
-			err = capzv1alpha3.AddToScheme(scheme)
-			if err != nil {
-				panic(err)
-			}
-			err = providerv1alpha1.AddToScheme(scheme)
-			if err != nil {
-				panic(err)
-			}
-			err = releasev1alpha1.AddToScheme(scheme)
-			if err != nil {
-				panic(err)
-			}
-
-			ctrlClient := fake.NewFakeClientWithScheme(scheme)
+			ctx := context.Background()
+			fakeK8sClient := unittest.FakeK8sClient()
+			ctrlClient := fakeK8sClient.CtrlClient()
 
 			// Create AzureMachinePool.
 			if tc.vmType != "" {
@@ -227,10 +191,22 @@ func TestMachinePoolCreateValidate(t *testing.T) {
 						},
 					},
 				}
-				err = ctrlClient.Create(context.Background(), amp)
+				err = ctrlClient.Create(ctx, amp)
 				if err != nil {
 					t.Fatal(err)
 				}
+			}
+
+			// Create default GiantSwarm organization.
+			organization := &securityv1alpha1.Organization{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "giantswarm",
+				},
+				Spec: securityv1alpha1.OrganizationSpec{},
+			}
+			err = ctrlClient.Create(ctx, organization)
+			if err != nil {
+				t.Fatal(err)
 			}
 
 			admit := &CreateValidator{
@@ -240,7 +216,7 @@ func TestMachinePoolCreateValidate(t *testing.T) {
 			}
 
 			// Run admission request to validate AzureConfig updates.
-			allowed, err := admit.Validate(context.Background(), getCreateAdmissionRequest(tc.machinePool))
+			allowed, err := admit.Validate(ctx, getCreateAdmissionRequest(tc.machinePool))
 
 			// Check if the error is the expected one.
 			switch {