+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Validate that the Organization label contains an existing Organization.
+
 ## [1.12.0] - 2020-10-27
 
 ### Added